Separating FCOnboardingData
